--- src/main/java/com/zerobase/moy/common/CommonResponse.java
@@ -1,24 +1,0 @@
-package com.zerobase.moy.common;
-
-// 예제 13.27
-public enum CommonResponse {
-
-  SUCCESS(0, "Success"), FAIL(-1, "Fail");
-
-  int code;
-  String msg;
-
-  CommonResponse(int code, String msg) {
-    this.code = code;
-    this.msg = msg;
-  }
-
-  public int getCode() {
-    return code;
-  }
-
-  public String getMsg() {
-    return msg;
-  }
-
-}
