+++ src/main/java/com/zerobase/moy/data/model/sign/LogoutResultDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.moy.data.model;
+package com.zerobase.moy.data.model.sign;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,12 +11,7 @@ import lombok.ToString;
 @ToString
 @Builder
 @Getter
-public class LogoutResultDto extends SignUpResultDto {
+public class LogoutResultDto {
 
-  private boolean success;
-
-  private int code;
-
-  private String msg;
   private String expiredToken;
 }
