+++ src/main/java/com/zerobase/moy/data/model/sign/TokenDto.java
@@ -1,9 +1,10 @@
-package com.zerobase.moy.data.model;
+package com.zerobase.moy.data.model.sign;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.ToString;
 
 @NoArgsConstructor
@@ -11,10 +12,9 @@ import lombok.ToString;
 @ToString
 @Builder
 @Getter
+@Setter
+public class TokenDto {
 
-public class SignUpRequestDto {
-
-  private String email;
-  private String password;
-  private String name;
+  private String atk;
+  private String rtk;
 }
